fix(KandelSower): Make smokeTest use the correct tokens
The base and quote arguments given to smokeTest were the MangroveTest.base and quote fields which were never initialized (because they’re not meaningful here).

It was probably a merge from master to develop with MGV v2 that went wrong.  And it wasn’t discovered because the base contract (MangroveTest) declared the now-missing values. And there isn’t a test for KandelSower that catches this...