ops(deploy): Deploy contracts to Mumbai (v0.0.2-3)
* fix(KandelSower): Make smokeTest use the correct tokens

The base and quote arguments given to smokeTest were the MangroveTest.base and quote fields which were never initialized (because they’re not meaningful here).

It was probably a merge from master to develop with MGV v2 that went wrong.  And it wasn’t discovered because the base contract (MangroveTest) declared the now-missing values. And there isn’t a test for KandelSower that catches this...

* feat(deploy): Update Mumbai full deployment script to current tokens and markets

The script only worked with AAVE tokens before. Now it handles AAVE tokens and Mangrove's own tokens separately. There's no price oracle for the latter, so the prices are hard-coded.

The script now opens these markets:

- AAVE tokens: WBTC/DAI, CRV/WBTC
- Own tokens: WMATIC/USDT, USDC/USDT, WETH/USDT

* ops(deploy): Deploy contracts to Mumbai (v0.0.2-3)

* fix(package.json): Remove wrong 'stableVersion'

This is known to cause issues